renamed sample program to from helloserver to podtatoserver
