--- hello-server/go.mod
@@ -1,5 +1,0 @@
-module github.com/keptn-sandbox/podtato-head/hello-server
-
-go 1.15
-
-require github.com/prometheus/client_golang v1.8.0
