+++ podtato-server/DockerfileV0.1.0
@@ -1,0 +1,43 @@
+
+FROM golang:alpine AS builder
+
+LABEL org.opencontainers.image.source https://github.com/cncf-podtato-head
+
+# Set necessary environment variables needed for our image
+ENV GO111MODULE=on \
+    CGO_ENABLED=0 \
+    GOOS=linux \
+    GOARCH=amd64
+
+# Move to working directory /build
+WORKDIR /build
+
+# Copy and download dependency using go mod
+# COPY go.mod .
+# COPY go.sum .
+# RUN go mod download
+
+# Copy the code into the container
+COPY . .
+
+# Build the application
+RUN go build -o podtatoserver ./podtatoserver.go
+
+# Move to /dist directory as the place for resulting binary folder
+WORKDIR /dist
+
+# Copy binary from build to main folder
+RUN cp /build/podtatoserver .
+
+# Build a small image
+FROM scratch
+
+COPY --from=builder /dist/podtatoserver /
+COPY ./static  /static
+COPY ./overview.html  /overview.html
+
+# Export necessary port
+EXPOSE 9000
+
+# Command to run when starting the container
+CMD ["/podtatoserver", "0.1.0"]
