+++ podtato-server/go.mod
@@ -1,0 +1,5 @@
+module github.com/cncf/podtato-head/podtato-server
+
+go 1.15
+
+require github.com/prometheus/client_golang v1.8.0
